+++ unit-test/openvibe-toolkit/urImportScenarioFromFileTest.cpp
@@ -51,7 +51,7 @@ int urImportScenarioFromFileTest(int argc, char* argv[])
 		// Test settings
 		OVT_ASSERT(scenario.getSettingCount() == simpleScenarioSettings.size(), "Imported scenario does not have the correct number of settings");
 
-		for (size_t settingIndex = 0; settingIndex < simpleScenarioSettings.size(); settingIndex += 1)
+		for (uint32_t settingIndex = 0; settingIndex < static_cast<uint32_t>(simpleScenarioSettings.size()); settingIndex += 1)
 		{
 			CIdentifier settingTypeId;
 			OVT_ASSERT(scenario.getSettingType(settingIndex, settingTypeId), "Cannot get setting type");
@@ -83,7 +83,7 @@ int urImportScenarioFromFileTest(int argc, char* argv[])
 		// Test inputs
 		OVT_ASSERT(scenario.getInputCount() == simpleScenarioInputs.size(), "Imported scenario has wrong number of inputs");
 
-		for (size_t inputIndex = 0; inputIndex < simpleScenarioInputs.size(); inputIndex += 1)
+		for (uint32_t inputIndex = 0; inputIndex < static_cast<uint32_t>(simpleScenarioInputs.size()); inputIndex += 1)
 		{
 			CIdentifier inputTypeId;
 			OVT_ASSERT(scenario.getInputType(inputIndex, inputTypeId), "Cannot get input type");
@@ -103,7 +103,7 @@ int urImportScenarioFromFileTest(int argc, char* argv[])
 		// Test outputs
 		OVT_ASSERT(scenario.getOutputCount() == simpleScenarioOutputs.size(), "Imported scenario has wrong number of outputs");
 
-		for (size_t outputIndex = 0; outputIndex < simpleScenarioOutputs.size(); outputIndex += 1)
+		for (uint32_t outputIndex = 0; outputIndex < static_cast<uint32_t>(simpleScenarioOutputs.size()); outputIndex += 1)
 		{
 			CIdentifier outputTypeId;
 			OVT_ASSERT(scenario.getOutputType(outputIndex, outputTypeId), "Cannot get output type");
